Cleaned up temporary allocation in IR compiler.
Virtual registers are now allocated for operands in the same way as for
instruction results. This is achieved by explicitly never allocating
registers for constants, but for everything else.

The memory allocation instruction has been replaced by a new temporary
value node in the IR. Passing a temporary node as an operand will
automatically assign a virtual register to it without the need for the old
memory allocation instruction.